LayoutTests/imported/w3c/web-platform-tests/preload/preload-dynamic-csp-expected.txt

CONSOLE MESSAGE: Refused to load http://localhost:8800/preload/resources/square.png because it does not appear in the img-src directive of the Content Security Policy.


PASS Makes sure that preload requests respect CSP directives that are added after the preload
